+++ source/bin.ts
@@ -1,8 +1,10 @@
 #!/usr/bin/env node
 import esrun from "./index.js"
 import { CliOption } from "./types/CliOption.js"
+import { Parameter } from "./types/Parameter.js"
 
 const { argv } = process
+const nodeOptionPrefix = "--node-"
 
 const argumentOptions: Record<string, CliOption> = {
 	"--watch": "watch",
@@ -14,12 +16,15 @@ const argumentOptions: Record<string, CliOption> = {
 	"--tsconfig": "tsconfig",
 }
 
-const options: Record<CliOption, boolean | string[] | undefined> = {
+const options: Record<CliOption, boolean | string[] | undefined> & {
+	node: Record<string, Parameter>
+} = {
 	watch: false,
 	inspect: false,
 	preserveConsole: false,
 	noFileConstants: false,
 	tsconfig: undefined,
+	node: {},
 }
 
 let argsOffset = 2
@@ -30,11 +35,14 @@ if (argv.length < argsOffset) {
 	process.exit(0)
 }
 
-while ((argument = argv[argsOffset]).startsWith("-")) {
-	const [command, parameters] = argument.split(":")
+while ((argument = argv[argsOffset]).startsWith('--')) {
+	const [command, parameters] = getCommandAndParameters(argument)
 
 	if (command in argumentOptions) {
-		options[argumentOptions[command]] = parameters ? parameters.split(",") : true
+		options[argumentOptions[command]] = parameters
+		argsOffset++
+	} else if (command.startsWith(nodeOptionPrefix)) {
+		options.node[command.slice(nodeOptionPrefix.length)] = parameters
 		argsOffset++
 	} else {
 		console.log(`Unknown option ${command}`)
@@ -54,7 +62,20 @@ esrun(argv[argsOffset], {
 	inspect: !!options.inspect,
 	preserveConsole: !!options.preserveConsole,
 	fileConstants: !options.noFileConstants,
+	nodeOptions: options.node,
 }).catch(error => {
 	console.error(error)
 	process.exit(1)
 })
+
+
+function getCommandAndParameters(argument: string): [string, Parameter] {
+	let colonIndex = argument.indexOf(':')
+	if (colonIndex == -1) colonIndex = Infinity
+	let equalIndex = argument.indexOf('=')
+	if (equalIndex == -1) equalIndex = Infinity
+	const separatorIndex = Math.min(colonIndex, equalIndex)
+
+	if (separatorIndex == Infinity) return [argument, true]
+	return [argument.slice(0, separatorIndex), argument.slice(separatorIndex + 1).split(',')]
+}