+++ source/types/Parameter.ts
@@ -1,0 +1,1 @@
+export type Parameter = string[] | true